+++ src/khoj/routers/api.py
@@ -19,6 +19,7 @@ from fastapi.responses import Response
 from starlette.authentication import has_required_scope, requires
 
 from khoj.configure import initialize_content
+from khoj.database import adapters
 from khoj.database.adapters import (
     AutomationAdapters,
     ConversationAdapters,
@@ -274,6 +275,38 @@ async def transcribe(
     # Return the spoken text
     content = json.dumps({"text": user_message})
     return Response(content=content, media_type="application/json", status_code=200)
+
+
+@api.patch("/user/name", status_code=200)
+@requires(["authenticated"])
+def set_user_name(
+    request: Request,
+    name: str,
+    client: Optional[str] = None,
+):
+    user = request.user.object
+
+    split_name = name.split(" ")
+
+    if len(split_name) > 2:
+        raise HTTPException(status_code=400, detail="Name must be in the format: Firstname Lastname")
+
+    if len(split_name) == 1:
+        first_name = split_name[0]
+        last_name = ""
+    else:
+        first_name, last_name = split_name[0], split_name[-1]
+
+    adapters.set_user_name(user, first_name, last_name)
+
+    update_telemetry_state(
+        request=request,
+        telemetry_type="api",
+        api="set_user_name",
+        client=client,
+    )
+
+    return {"status": "ok"}
 
 
 async def extract_references_and_questions(
